pyproject.toml
[project]
name = "gui-react-agent"
version = "0.1.0"
description = "GUI ReAct Agent - 基于纯视觉感知的GUI自动化系统"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # 核心依赖
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    # 环境变量管理
    "python-dotenv>=1.0.0",
    # 浏览器自动化
    "playwright>=1.40.0",
    # 可选：桌面自动化
    # "pyautogui>=0.9.54",
    # 可选：Gradio UI
    "gradio>=4.0.0",
]

[project.optional-dependencies]
desktop = [
    "pyautogui>=0.9.54",
    "pillow>=10.0.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
gui-agent = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "examples"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.black]
line-length = 100
